analyzer/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
requests
python-dotenv
tiktoken
